agent/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
supabase==2.5.0
anthropic==0.34.0
python-dotenv==1.0.0
pydantic==2.7.0
pydantic-settings==2.3.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
httpx==0.27.0
pyyaml==6.0.1
apscheduler==3.10.4
pytz==2024.1
